update: 試着する画面のui
