+++ locales/translations/ja.json
@@ -227,7 +227,9 @@
       "title": "生成中...",
       "description": "あなたにぴったりの服を見つけています。\nしばらくお待ちください。"
     },
-    "error": "エラーが発生しました。もう一度お試しください。"
+    "error": "エラーが発生しました。もう一度お試しください。",
+    "tops": "トップス",
+    "bottoms": "ボトムス"
   },
   "discover": {
     "title": "見つける",
